requirements.txt
discord.py
python-dotenv
groq
feedparser
aiohttp